pyproject.toml
[tool.poetry]
name = "AnimeGrok"
version = "0.1.0"
description = "AnimeGrok"
authors = ["Artur R <ra072016@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
django-mptt = "^0.14.0"
django-ckeditor = "^6.5.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.7"
pandas = "^1.5.3"
django-debug-toolbar = "^4.0.0"
django-allauth = "^0.54.0"
httpie = "^3.2.1"
pygments = "^2.15.0"
django-filter = "^23.1"
celery = {extras = ["redis"], version = "^5.2.7"}
drf-yasg = "^1.21.5"
python-slugify = {extras = ["unidecode"], version = "^8.0.1"}
django-recaptcha3 = "^0.4.0"
django-modeltranslation = "^0.18.9"
django-redis = "^5.2.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
djipsum = "^1.1.5"
coverage = "^7.2.7"
mock = "^5.0.2"
django-factory-boy = "^1.0.0"


[build-system]
requires = ["poetry-core==1.3.2"]
build-backend = "poetry.core.masonry.api"
